updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -36,10 +36,10 @@ with app.app_context():
 # get called. What it returns is what is shown as the web page
 @app.route('/index')
 def index():
-    #get user from database
-    a_user =  db.session.query(User).filter_by(email='eevensen@uncc.edu').one()
+    if session.get('user'):
+        return render_template('index.html', user=session['user'])
 
-    return render_template('index.html', user = a_user)
+    return render_template('index.html')
 
 @app.route('/notes')
 def get_notes():
@@ -60,55 +60,60 @@ def get_note(note_id):
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
 
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
 
-        import datetime
-        today = datetime.datetime.now()
-        today = today.strftime("%m-%d-%Y")
+            import datetime
+            today = datetime.datetime.now()
+            today = today.strftime("%m-%d-%Y")
 
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user']) 
     else:
-        a_user = db.session.query(User).filter_by(email='eevensen@uncc.edu').one()
-        return render_template('new.html', user=a_user)  
+        return redirect(url_for('login')) 
 
 app.run(host=os.getenv('IP', '127.0.0.1'),port=int(os.getenv('PORT', 5000)),debug=True)
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    if request.method == 'POST':
-        title = request.form['title']
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
 
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
 
-        db.session.add(note)
+            db.session.add(note)
+            db.session.commit()
+
+            return redirect(url_for('get_notes'))
+        else:
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note=my_note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+@app.route('/notes/delete/<note_id>', methods=['POST'])
+def delete_note(note_id):
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
         db.session.commit()
 
         return redirect(url_for('get_notes'))
     else:
-
-        a_user = db.session.query(User).filter_by(email='eevensen@uncc.edu').one()
-
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
-
-@app.route('/notes/delete/<note_id>', methods=['POST'])
-def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
